requirements.txt
fastapi
uvicorn
openai
faiss-cpu
tiktoken
numpy
python-dotenv